Multi Hash Update
Added ability to specify multiple hashes by simply typing them in. E.g. "md5 sha 1 sha224 sha256sha512 sha3256 s  h  a  k  e  1  2  8     "
This works by removing all spaces making the above text "md5sha1sha224sha256sha512sha3256shake128" then checking to see if a specific string is in the text, for example "md5" is in the string so it runs the MD5 hash algorithm and "shake128" runs the SHAKE128 hash algorithm.

Simplified the hashFile and the stringFile to make them work with the multi-hashing option.

Considered adding a hashing progress text, but that became too complicated and would require the entire code to be restructured. Maybe later.